fix: pages src 하위 폴더로 변경
